frontend/requirements.txt
streamlit==1.29.0
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.1.4